Add spinPath

+++ etc/StoringMethodForData/app/src/main/java/com/mutecsoft/storingmethodfordata/FilesActivity.java
@@ -5,94 +5,73 @@ import android.content.pm.PackageManager;
 import android.os.Bundle;
 import android.os.Environment;
 import android.view.View;
+import android.widget.AdapterView;
+import android.widget.ArrayAdapter;
 import android.widget.Button;
 import android.widget.EditText;
+import android.widget.Spinner;
 import android.widget.Toast;
 
 import java.io.File;
 import java.io.FileInputStream;
 import java.io.FileOutputStream;
-import java.security.Permission;
 
 /**
  * Created by mutecsoft on 2016-06-15.
  */
 public class FilesActivity extends Activity {
 
+    private static final int PERMISSION_REQUEST_CODE = 999;
     private static final String FILE_NAME = "Memo.txt";
+
+    String[] permissions = {"android.permission.READ_EXTERNAL_STORAGE", "android.permission.WRITE_EXTERNAL_STORAGE"};
 
     Button btnRead, btnSave, btnDelete;
     EditText et;
-
+    Spinner spinPath;
+    String selectedPath;
     File file;
-
-    String[] permissions = {"android.permission.READ_EXTERNAL_STORAGE", "android.permission.WRITE_EXTERNAL_STORAGE"};
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_files);
 
+        if(checkSelfPermission("android.permission.READ_EXTERNAL_STORAGE") == PackageManager.PERMISSION_DENIED || checkSelfPermission("android.permission.WRITE_EXTERNAL_STORAGE") == PackageManager.PERMISSION_DENIED){
+            requestPermissions(permissions, PERMISSION_REQUEST_CODE);
+        }
+
         btnRead = (Button) findViewById(R.id.btnRead);
         btnSave = (Button) findViewById(R.id.btnSave);
         btnDelete = (Button) findViewById(R.id.btnDelete);
         et = (EditText) findViewById(R.id.et);
+        spinPath = (Spinner) findViewById(R.id.spinPath);
 
-        String[] permissions = {"android.permission.READ_EXTERNAL_STORAGE", "android.permission.WRITE_EXTERNAL_STORAGE"};
+        String[] pathList = {"App", "Download"};
+        ArrayAdapter<String> adapter = new ArrayAdapter<String>(this, android.R.layout.simple_expandable_list_item_1, android.R.id.text1, pathList);
+        spinPath.setAdapter(adapter);
 
-        if(checkSelfPermission("android.permission.READ_EXTERNAL_STORAGE") == PackageManager.PERMISSION_DENIED || checkSelfPermission("android.permission.WRITE_EXTERNAL_STORAGE") == PackageManager.PERMISSION_DENIED){
-            requestPermissions(permissions, 0);
-        }
+        spinPath.setOnItemSelectedListener(new AdapterView.OnItemSelectedListener() {
+            @Override
+            public void onItemSelected(AdapterView<?> parent, View view, int position, long id) {
+                String path = spinPath.getSelectedItem().toString();
 
-        // 파일 생성및 데이터 저장
-        try {
-            FileOutputStream fos = openFileOutput("file_name", MODE_PRIVATE);
-            fos.write("strData".getBytes());
-        }catch (Exception e){
-            e.printStackTrace();
-        }
-        // 파일 데이터 읽기
-        try {
-            FileInputStream fis = openFileInput("file_name");
-            byte[] bytesData = new byte[fis.available()];
-            while (fis.read(bytesData) != -1) {}
-            String strData = new String(bytesData);
-        }catch (Exception e){
-            e.printStackTrace();
-        }
-
-        try {
-            file = new File(Environment.getExternalStoragePublicDirectory(Environment.DIRECTORY_DOWNLOADS).toString()+"/poiu.txt");
-            if (file!=null && !file.exists()){
-                try {
-                    file.createNewFile();
-                }catch (Exception e){
-                    e.printStackTrace();
+                switch (path){
+                    case "App" :
+                        selectedPath = "App";
+                        break;
+                    case "Download" :
+                        selectedPath = Environment.getExternalStoragePublicDirectory(Environment.DIRECTORY_DOWNLOADS).toString()+"/"+FILE_NAME;
+                        file = new File(selectedPath);
+                        break;
                 }
             }
-            Toast.makeText(getApplicationContext(), Environment.getExternalStoragePublicDirectory(Environment.DIRECTORY_DOWNLOADS).toString(), Toast.LENGTH_SHORT).show();
-            Toast.makeText(getApplicationContext(), "success", Toast.LENGTH_SHORT).show();
-        }catch (Exception e){
-            e.printStackTrace();
-            Toast.makeText(getApplicationContext(), "fail", Toast.LENGTH_SHORT).show();
-        }
 
-        if (file!=null && !file.exists()){
-            try {
-                file.createNewFile();
-                Toast.makeText(getApplicationContext(), "new", Toast.LENGTH_SHORT).show();
-            }catch (Exception e){
-                e.printStackTrace();
-                Toast.makeText(getApplicationContext(), e.toString(), Toast.LENGTH_SHORT).show();
+            @Override
+            public void onNothingSelected(AdapterView<?> parent) {
+
             }
-        }
-        try {
-            FileOutputStream fos = new FileOutputStream(file);
-            fos.write("아니바그아고야".getBytes());
-            fos.close();
-        }catch (Exception e){
-            e.printStackTrace();
-        }
+        });
 
 
 
@@ -110,34 +89,61 @@ public class FilesActivity extends Activity {
             switch (v.getId()){
                 case R.id.btnRead :
                     try {
-                        FileInputStream fis = openFileInput(FILE_NAME);
-
-                        byte[] memoData = new byte[fis.available()];
-                        while (fis.read(memoData) != -1){
-
+                        FileInputStream fis = null;
+                        if (selectedPath.equals("App")){
+                            fis = openFileInput(FILE_NAME);
+                        }else{
+                            fis = new FileInputStream(file);
                         }
+                        byte[] memoData = new byte[fis.available()];
+                        while (fis.read(memoData) != -1){}
                         et.setText(new String(memoData));
-                        Toast.makeText(getApplicationContext(), "Completed read", Toast.LENGTH_LONG).show();
+                        Toast.makeText(getApplicationContext(), "Completed read", Toast.LENGTH_SHORT).show();
                     }catch (Exception e){
                         e.printStackTrace();
                     }
                     break;
                 case R.id.btnSave :
-                    FileOutputStream fos = null;
                     try {
-                        fos = openFileOutput(FILE_NAME, MODE_PRIVATE);
+                        FileOutputStream fos = null;
+                        if (selectedPath.equals("App")){
+                            fos = openFileOutput(FILE_NAME, MODE_PRIVATE);
+                        }else{
+                            fos = new FileOutputStream(file);
+                        }
                         fos.write(et.getText().toString().getBytes());
-                        Toast.makeText(getApplicationContext(), "Completed save", Toast.LENGTH_LONG).show();
+                        Toast.makeText(getApplicationContext(), "Completed save", Toast.LENGTH_SHORT).show();
                     }catch (Exception e){
                         e.printStackTrace();
                     }
                     break;
                 case R.id.btnDelete :
-                    if (deleteFile(FILE_NAME)) {
-                        Toast.makeText(getApplicationContext(), "Completed delete", Toast.LENGTH_LONG).show();
-                        et.setText("");
+                    try {
+                        if (selectedPath.equals("App")){
+                            if (deleteFile(FILE_NAME)) {
+                                Toast.makeText(getApplicationContext(), "Completed delete", Toast.LENGTH_SHORT).show();
+                                et.setText("");
+                            }
+                        }else{
+                            if (file.delete()) {
+                                Toast.makeText(getApplicationContext(), "Completed delete", Toast.LENGTH_SHORT).show();
+                                et.setText("");
+                            }
+                        }
+                    }catch (Exception e){
+                        e.printStackTrace();
                     }
+            }
+        }
+    }
 
+    @Override
+    public void onRequestPermissionsResult(int requestCode, String[] permissions, int[] grantResults) {
+        super.onRequestPermissionsResult(requestCode, permissions, grantResults);
+        if (requestCode == PERMISSION_REQUEST_CODE){
+            if(checkSelfPermission("android.permission.READ_EXTERNAL_STORAGE") == PackageManager.PERMISSION_DENIED || checkSelfPermission("android.permission.WRITE_EXTERNAL_STORAGE") == PackageManager.PERMISSION_DENIED){
+                Toast.makeText(getApplicationContext(), "PERMISSION_DENIED", Toast.LENGTH_SHORT).show();
+                finish();
             }
         }
     }
